Shib: API test and javadoc cleanup (no-op) #3025
