+++ src/test/java/edu/harvard/iq/dataverse/api/AdminIT.java
@@ -2,13 +2,11 @@ package edu.harvard.iq.dataverse.api;
 
 import com.jayway.restassured.RestAssured;
 import com.jayway.restassured.response.Response;
-import static edu.harvard.iq.dataverse.api.UtilIT.API_TOKEN_HTTP_HEADER;
 import static javax.ws.rs.core.Response.Status.FORBIDDEN;
 import static javax.ws.rs.core.Response.Status.OK;
 import static javax.ws.rs.core.Response.Status.BAD_REQUEST;
 import org.junit.Test;
 import org.junit.BeforeClass;
-import static com.jayway.restassured.RestAssured.given;
 import java.util.UUID;
 import static org.hamcrest.CoreMatchers.equalTo;
 import org.junit.AfterClass;
@@ -52,15 +50,15 @@ public class AdminIT {
 
     @Test
     public void testListAuthenticatedUsers() throws Exception {
-        Response anon = listAuthenticatedUsers("");
+        Response anon = UtilIT.listAuthenticatedUsers("");
         anon.prettyPrint();
         anon.then().assertThat().statusCode(FORBIDDEN.getStatusCode());
 
-        Response nonSuperuser = listAuthenticatedUsers(nonSuperuserApiToken);
+        Response nonSuperuser = UtilIT.listAuthenticatedUsers(nonSuperuserApiToken);
         nonSuperuser.prettyPrint();
         nonSuperuser.then().assertThat().statusCode(FORBIDDEN.getStatusCode());
 
-        Response superuser = listAuthenticatedUsers(superuserApiToken);
+        Response superuser = UtilIT.listAuthenticatedUsers(superuserApiToken);
         superuser.prettyPrint();
         superuser.then().assertThat().statusCode(OK.getStatusCode());
 
@@ -73,11 +71,11 @@ public class AdminIT {
         String newEmailAddressToUse = "builtin2shib." + UUID.randomUUID().toString().substring(0, 8) + "@mailinator.com";
         String data = emailOfUserToConvert + ":" + password + ":" + newEmailAddressToUse;
 
-        Response builtinToShibAnon = migrateBuiltinToShib(data, "");
+        Response builtinToShibAnon = UtilIT.migrateBuiltinToShib(data, "");
         builtinToShibAnon.prettyPrint();
         builtinToShibAnon.then().assertThat().statusCode(FORBIDDEN.getStatusCode());
 
-        Response makeShibUser = migrateBuiltinToShib(data, superuserApiToken);
+        Response makeShibUser = UtilIT.migrateBuiltinToShib(data, superuserApiToken);
         makeShibUser.prettyPrint();
         makeShibUser.then().assertThat()
                 .statusCode(OK.getStatusCode())
@@ -89,15 +87,15 @@ public class AdminIT {
          * the Shib user has an invalid email address:
          * https://github.com/IQSS/dataverse/issues/2998
          */
-        Response shibToBuiltinAnon = migrateShibToBuiltin(Long.MAX_VALUE, "", "");
+        Response shibToBuiltinAnon = UtilIT.migrateShibToBuiltin(Long.MAX_VALUE, "", "");
         shibToBuiltinAnon.prettyPrint();
         shibToBuiltinAnon.then().assertThat().statusCode(FORBIDDEN.getStatusCode());
 
-        Response nonSuperuser = migrateShibToBuiltin(Long.MAX_VALUE, "", "");
+        Response nonSuperuser = UtilIT.migrateShibToBuiltin(Long.MAX_VALUE, "", "");
         nonSuperuser.prettyPrint();
         nonSuperuser.then().assertThat().statusCode(FORBIDDEN.getStatusCode());
 
-        Response infoOfUserToConvert = getAuthenticatedUser(usernameOfUserToConvert, superuserApiToken);
+        Response infoOfUserToConvert = UtilIT.getAuthenticatedUser(usernameOfUserToConvert, superuserApiToken);
         infoOfUserToConvert.prettyPrint();
         /**
          * https://github.com/IQSS/dataverse/issues/2418 is cropping up again in
@@ -107,17 +105,17 @@ public class AdminIT {
         idOfUserToConvert = infoOfUserToConvert.body().jsonPath().getLong("data.id");
 
         String invalidEmailAddress = "invalidEmailAddress";
-        Response invalidEmailFail = migrateShibToBuiltin(idOfUserToConvert, invalidEmailAddress, superuserApiToken);
+        Response invalidEmailFail = UtilIT.migrateShibToBuiltin(idOfUserToConvert, invalidEmailAddress, superuserApiToken);
         invalidEmailFail.prettyPrint();
         invalidEmailFail.then().assertThat().statusCode(BAD_REQUEST.getStatusCode());
 
         String existingEmailAddress = "dataverse@mailinator.com";
-        Response existingEmailFail = migrateShibToBuiltin(idOfUserToConvert, existingEmailAddress, superuserApiToken);
+        Response existingEmailFail = UtilIT.migrateShibToBuiltin(idOfUserToConvert, existingEmailAddress, superuserApiToken);
         existingEmailFail.prettyPrint();
         existingEmailFail.then().assertThat().statusCode(BAD_REQUEST.getStatusCode());
 
         String newEmailAddress = UUID.randomUUID().toString().substring(0, 8) + "@mailinator.com";
-        Response shouldWork = migrateShibToBuiltin(idOfUserToConvert, newEmailAddress, superuserApiToken);
+        Response shouldWork = UtilIT.migrateShibToBuiltin(idOfUserToConvert, newEmailAddress, superuserApiToken);
         shouldWork.prettyPrint();
         shouldWork.then().assertThat().statusCode(OK.getStatusCode());
 
@@ -140,39 +138,6 @@ public class AdminIT {
         Response deleteSuperuser = UtilIT.deleteUser(superuserUsername);
         assertEquals(200, deleteSuperuser.getStatusCode());
 
-    }
-
-    /**
-     * @todo: Once all these methods have stabilized, move to UtilIT.java
-     */
-    static Response listAuthenticatedUsers(String apiToken) {
-        Response response = given()
-                .header(API_TOKEN_HTTP_HEADER, apiToken)
-                .get("/api/admin/authenticatedUsers");
-        return response;
-    }
-
-    static Response getAuthenticatedUser(String userIdentifier, String apiToken) {
-        Response response = given()
-                .header(API_TOKEN_HTTP_HEADER, apiToken)
-                .get("/api/admin/authenticatedUsers/" + userIdentifier);
-        return response;
-    }
-
-    static Response migrateShibToBuiltin(Long userIdToConvert, String newEmailAddress, String apiToken) {
-        Response response = given()
-                .header(API_TOKEN_HTTP_HEADER, apiToken)
-                .body(newEmailAddress)
-                .put("/api/admin/authenticatedUsers/id/" + userIdToConvert + "/convertShibToBuiltIn");
-        return response;
-    }
-
-    static Response migrateBuiltinToShib(String data, String apiToken) {
-        Response response = given()
-                .header(API_TOKEN_HTTP_HEADER, apiToken)
-                .body(data)
-                .put("/api/admin/authenticatedUsers/convert/builtin2shib");
-        return response;
     }
 
 }
