+++ src/test/java/edu/harvard/iq/dataverse/api/UtilIT.java
@@ -381,6 +381,36 @@ public class UtilIT {
         return response;
     }
 
+    static Response listAuthenticatedUsers(String apiToken) {
+        Response response = given()
+                .header(API_TOKEN_HTTP_HEADER, apiToken)
+                .get("/api/admin/authenticatedUsers");
+        return response;
+    }
+
+    static Response getAuthenticatedUser(String userIdentifier, String apiToken) {
+        Response response = given()
+                .header(API_TOKEN_HTTP_HEADER, apiToken)
+                .get("/api/admin/authenticatedUsers/" + userIdentifier);
+        return response;
+    }
+
+    static Response migrateShibToBuiltin(Long userIdToConvert, String newEmailAddress, String apiToken) {
+        Response response = given()
+                .header(API_TOKEN_HTTP_HEADER, apiToken)
+                .body(newEmailAddress)
+                .put("/api/admin/authenticatedUsers/id/" + userIdToConvert + "/convertShibToBuiltIn");
+        return response;
+    }
+
+    static Response migrateBuiltinToShib(String data, String apiToken) {
+        Response response = given()
+                .header(API_TOKEN_HTTP_HEADER, apiToken)
+                .body(data)
+                .put("/api/admin/authenticatedUsers/convert/builtin2shib");
+        return response;
+    }
+
     @Test
     public void testGetFileIdFromSwordStatementWithNoFiles() {
         String swordStatementWithNoFiles = "<feed xmlns=\"http://www.w3.org/2005/Atom\">\n"
